move table options underneath spec pkg
