+++ plugins/source/aws/client/spec/tableoptions/cloudtrail.go
