+++ plugins/source/aws/client/spec/tableoptions/cloudtrail_test.go
