+++ plugins/source/aws/client/spec/tableoptions/inspector2.go
