+++ plugins/source/aws/client/spec/tableoptions/inspector2_test.go
