+++ plugins/source/aws/client/spec/tableoptions/securityhub.go
