+++ plugins/source/aws/client/spec/tableoptions/securityhub_test.go
