+++ plugins/source/aws/client/spec/tableoptions/table_options.go
