+++ plugins/source/aws/client/spec/tableoptions/table_options_test.go
